Removed dependency on JSON.NET
Replaced JSON.NET with System.Runtime.Serialization.Json to
remove the external dependency.
